parser: Create the `rules` and `tokens` modules.
