+++ source/parser/grammar/mod.rs
@@ -31,65 +31,6 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-mod ast;
-
-use nom::digit;
-use nom::IResult::Done;
-use std::str;
-use std::str::FromStr;
-
-named!(
-    i64_digit<i64>,
-    map_res!(
-        map_res!(
-            digit,
-            str::from_utf8
-        ),
-        FromStr::from_str
-    )
-);
-
-named!(
-    expr<ast::Addition>,
-    chain!(
-        left: i64_digit ~
-        tag!("+") ~
-        right: i64_digit,
-        || { ast::Addition { a: ast::Term { t: left }, b: ast::Term { t: right } } }
-    )
-);
-
-pub fn root(input: &[u8]) -> ast::Addition {
-    match expr(input) {
-        Done(_, ast) => ast,
-        _ => panic!("Youhouuu")
-    }
-}
-
-
-#[cfg(test)]
-mod tests {
-    use super::ast::Addition;
-    use super::ast::Term;
-    use nom::IResult::Done;
-    use super::i64_digit;
-    use super::expr;
-
-    #[test]
-    fn case_i64_digit() {
-        assert_eq!(
-            i64_digit(b"42"),
-            Done(&b""[..], 42)
-        );
-    }
-
-    #[test]
-    fn case_expr() {
-        assert_eq!(
-            expr(b"1+2"),
-            Done(
-                &b""[..], Addition { a: Term { t: 1 }, b: Term { t: 2 } }
-            )
-        );
-    }
-}
+pub mod ast;
+pub mod rules;
+pub mod tokens;
